feat: Add support for fetching CephFS mirroring status
